Do not allow yamux to log directly to stderr
By default, yamux logs directly to stderr. The go-plugin framework
currently uses a default yamux client when initializing a RPC client.
This results in "uncapturable" log lines appearing on stderr when
communicating with a plugin, for example:

2021/03/16 23:13:31 [ERR] yamux: Failed to write header: write unix @->/tmp/plugin383370083: write: broken pipe
2021/03/16 23:15:41 [ERR] yamux: Failed to write header: write unix @->/tmp/plugin806433500: use of closed network connection

This change gives the RPC client's configured logger to yamux to write
to and allows yamux to attempt to infer the log level based on
indicators such as "[ERR]" which isn't a bad assumption based on how
yamux's log lines look.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>